* upstream/master: (384 commits)
  Update docs
  Fix path to link.html in new tests
  Don't autofollow hint when unfiltering w/ rapid
  bdd: Wait a bit in "No crash should happen" step
  tox: Update pytest-mock to 1.0.0
  Fix lint
  Implement utils.is_special_key.
  Don't show keyhint if there are no hints.
  Regenerate authors
  Add invalid url to 'test_set_hover_url_encoded'
  Update changelog
  Try to fix broken yelo link in README
  Fix crash when hovering over an invalid URL
  Use ', '.join instead of str.join(', ', ...).
  Update command completions when keyconf changes.
  Show special keys last in command completion.
  Clean up command_binding_completion.
  Update docs
  Update changelog
  Set Qt.RichText textFormat for KeyHintView
  ...

 Conflicts:
	qutebrowser/browser/hints.py

+++ qutebrowser/browser/adblock.py
@@ -116,6 +116,7 @@ class HostBlocker:
             self._local_hosts_file = None
         else:
             self._local_hosts_file = os.path.join(data_dir, 'blocked-hosts')
+        self.on_config_changed()
 
         config_dir = standarddir.config()
         if config_dir is None:
@@ -176,7 +177,8 @@ class HostBlocker:
                 message.info('current',
                              "Run :adblock-update to get adblock lists.")
 
-    @cmdutils.register(instance='host-blocker', win_id='win_id')
+    @cmdutils.register(instance='host-blocker')
+    @cmdutils.argument('win_id', win_id=True)
     def adblock_update(self, win_id):
         """Update the adblock block lists.
 
@@ -273,11 +275,13 @@ class HostBlocker:
     def on_config_changed(self):
         """Update files when the config changed."""
         urls = config.get('content', 'host-block-lists')
-        if urls is None:
+        if urls is None and self._local_hosts_file is not None:
             try:
                 os.remove(self._local_hosts_file)
-            except OSError:
-                log.misc.exception("Failed to delete hosts file.")
+            except FileNotFoundError:
+                pass
+            except OSError as e:
+                log.misc.exception("Failed to delete hosts file: {}".format(e))
 
     def on_download_finished(self, download):
         """Check if all downloads are finished and if so, trigger reading.
